requirements.txt
streamlit==1.41.1
groq===0.13.1
langchain_community==0.3.13
langchain-groq==0.2.2
langchain_core==0.3.28
llama-parse==0.5.18
python-dotenv==1.0.1
streamlit-chat==0.1.1
notebook==7.3.2
langchain==0.3.13
loguru==0.7.2
sentence_transformers==3.3.1
langgraph==0.2.53
langgraph-checkpoint-sqlite==2.0.1
langgraph-checkpoint==2.0.5
faiss-cpu==1.9.0.post1
transformers==4.47.1
duckduckgo-search==7.0.2
tavily-python==0.5.0
tiktoken==0.8.0
langchain-openai==0.2.14